refactor use_cases into multiple files
